refactor: 위치 계산 로직 분리
